feat: Complete PID class and project
This commit completes the PID class which is essentially the entire project. The initial tau values were found by tweaking the values given in the lesson. The PID class is a simple implementation of a PID controller which updates the P, I, and D errors based on the current crosstrack error. This error is then used to calculate the new steering angle value.